notification/requirements.txt
Django==4.2.26
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
certifi==2024.2.2
